+++ activestorage/lib/active_storage/attached/one.rb
@@ -63,6 +63,16 @@ module ActiveStorage
       record.public_send("#{name}")
     end
 
+    # Calls attach, and raises an exception if the record was meant to be saved but at least one of the validations failed.
+    #
+    #   person.avatar.attach!(params[:avatar]) # ActionDispatch::Http::UploadedFile object
+    #   person.avatar.attach!(params[:signed_blob_id]) # Signed reference to blob from direct upload
+    #   person.avatar.attach!(io: File.open("/path/to/face.jpg"), filename: "face.jpg", content_type: "image/jpeg")
+    #   person.avatar.attach!(avatar_blob) # ActiveStorage::Blob object
+    def attach!(attachable)
+      attach(attachable) || raise(RecordNotSaved.new("Failed to save the record", record))
+    end
+
     # Returns +true+ if an attachment has been made.
     #
     #   class User < ApplicationRecord