+++ activestorage/CHANGELOG.md
@@ -31,3 +31,7 @@
     *Edouard-chin*
 
 Please check [8-0-stable](https://github.com/rails/rails/blob/8-0-stable/activestorage/CHANGELOG.md) for previous changes.
+
+*   Implement `attach!`. This method raises an exception if the attachment was not saved.
+
+    *Quentin de Metz*